requirements.txt
Django==3.2.18
gunicorn==20.1.0
djangorestframework==3.14.0
